--- a/src/Benchmark/Program.cs
+++ b/src/Benchmark/Program.cs
@@ -1,4 +1,8 @@
+using System.Linq;
+using System.Threading.Tasks;
 using BenchmarkDotNet.Running;
+using JustEat.StatsD;
+using JustEat.StatsD.EndpointLookups;
 
 namespace Benchmark
 {
@@ -6,10 +10,33 @@ namespace Benchmark
     {
         internal static void Main(string[] args)
         {
-            BenchmarkRunner.Run<StatSendingBenchmark>();
-            BenchmarkRunner.Run<FormatterBenchmark>();
+            var config = new StatsDConfiguration
+            {
+                Host = "127.0.0.1",
+            };
+
+            var endpointSource1 = EndpointParser.MakeEndPointSource(
+                config.Host,
+                config.Port,
+                config.DnsLookupInterval);
+
+            var endpointSource2 = EndpointParser.MakeEndPointSource(
+                config.Host,
+                config.Port + 2,
+                config.DnsLookupInterval);
+
+            var wtf = new MilisecondSwitcher(endpointSource1, endpointSource2);
+
+            var pooledTransport = new PooledUdpTransport(wtf);
+
+            Parallel.For(0, 10000, _ => pooledTransport.Send("hello.world.yo"));
+
+            pooledTransport.Dispose();
+
+            //BenchmarkRunner.Run<StatSendingBenchmark>();
+            //BenchmarkRunner.Run<FormatterBenchmark>();
             BenchmarkRunner.Run<UdpTransportBenchmark>();
-            BenchmarkRunner.Run<UdpStatSendingBenchmark>();
+            //BenchmarkRunner.Run<UdpStatSendingBenchmark>();
         }
     }
 }