--- a/src/JustEat.StatsD.Tests/WhenUsingPooledUdpTransport.cs
+++ b/src/JustEat.StatsD.Tests/WhenUsingPooledUdpTransport.cs
@@ -1,4 +1,8 @@
+using System;
+using System.Net;
+using System.Threading;
 using System.Threading.Tasks;
+using JustEat.StatsD.EndpointLookups;
 using Xunit;
 
 namespace JustEat.StatsD
@@ -47,6 +51,45 @@ namespace JustEat.StatsD
                     target.Send("mycustommetric");
                 });
             }
+        }
+
+        private class MilisecondSwitcher : IPEndPointSource
+        {
+            private readonly IPEndPointSource _endpointSource1;
+            private readonly IPEndPointSource _endpointSource2;
+
+            public MilisecondSwitcher(IPEndPointSource endpointSource1, IPEndPointSource endpointSource2)
+            {
+                _endpointSource1 = endpointSource1;
+                _endpointSource2 = endpointSource2;
+            }
+
+            public IPEndPoint GetEndpoint()
+            {
+                return DateTime.Now.Millisecond % 2 == 0 ?
+                    _endpointSource1.GetEndpoint() :
+                    _endpointSource2.GetEndpoint();
+            }
+        }
+
+        [Fact]
+        public static void MultipleMetricsCanBeSentWithoutAnExceptionBeingThrownParallel_WithChangingEndpoints()
+        {
+            // Arrange
+            var endPointSource1 = EndpointLookups.EndpointParser.MakeEndPointSource("127.0.0.1", 8125, null);
+            var endPointSource2 = EndpointLookups.EndpointParser.MakeEndPointSource("127.0.0.1", 8126, null);
+
+            var endPointSource = new MilisecondSwitcher(endPointSource1, endPointSource2);
+
+            using (var target = new PooledUdpTransport(endPointSource))
+            {
+                Parallel.For(0, 100_000, _ =>
+                {
+                    target.Send("mycustommetric");
+                });
+            }
+
+            Thread.Sleep(5000);
         }
     }
 }