--- a/src/JustEat.StatsD/ConnectedSocketPool.cs
+++ b/src/JustEat.StatsD/ConnectedSocketPool.cs
@@ -26,7 +26,7 @@ namespace JustEat.StatsD
                     }
                     catch
                     {
-                        socket?.Dispose();
+                        socket.Dispose();
                         throw;
                     }
                 });