@dv00d00 2018-07-19
I was able to reproduce test failures on my mac, created a ticket in corefx (https://github.com/dotnet/corefx/issues/31206)