@martincostello 2018-07-20
The failed test could just maybe be a manifestation of differences in OS behaviours a bit like how we have to have this code:

https://github.com/justeat/JustEat.StatsD/blob/554633f110b6e9292a5d69055c171e9a0ebc32da/src/JustEat.StatsD/UdpTransport.cs#L52-L56