@dv00d00 2018-07-24
I've fixed Unix tests by binding loopback ports to a UDP socket as seen in UdpListeners.cs